// packages/core-components/src/components/card/enum.ts
/**
 * Card Enum for CSS Classes
 */
export enum CssClass {
    Pattern = "osui-card",
    Modifier = "osui-card--is-",
    HeaderOverlay = "osui-card--is-overlay-",
    HorizontalAlignment = "osui-card--is-horizontal",
    VerticalAlignment = "osui-card--is-vertical",
    IsReverseColumn = "osui-card--is-reverse-column",
    ShowHeader = "osui-card--is-show-header",
    ShowContent = "osui-card--is-show-content",
    ShowFooter = "osui-card--is-show-footer",
    BackgroundCover = "osui-card--is-background-cover",
}

/**
 * Card CSS Variables
 */
export enum CssCustomProperties {
    BackgroundColor = "--osui-card-background-color",
    Border = "--osui-card-border-radius",
    ImageURL = "--osui-card-image-url",
    Position = "--osui-card-background-postion",
}

/**
 * Card Properties
 */
export enum Properties {
    Alignment = "alignment",
    Border = "border",
    ReverseColumns = "reverseColumns",
    ShowContent = "showContent",
    ShowFooter = "showFooter",
    ShowHeader = "showHeader",
    BackgroundColor = "backgroundColor",
}